Create README.md
Hi everyone, im a programming student and I wanted to keep track of the times I die in this game. I didn´t found any mod that allows me to do that so i decided to build a little program in C# that allows me to count it. This is my first "mod" and it is totally external to the game (you have to press a button every time you die) and I want to trigger the count to the death inside the game but I dont really know how to modify ingame code, If you know how to do it please contact me and we can do it together. The program allows you to add a death, change it to a specific value and of course save it on a file to resume it if you close the program, you don´t need to memorize it.